logger: convert: fix variable type for negative value
time_precision variable can be used as -1

Fixes commit ff9343aa4a59 ("logger: convert: Fix compile time error with
newer toolchain")

+++ tools/logger/convert.c
@@ -532,7 +532,7 @@ static void print_entry_params(const struct log_entry_header *dma_log,
 				entry->header.line_idx);
 	} else {
 		if (time_precision >= 0) {
-			const unsigned int ts_width = timestamp_width(time_precision);
+			const uint8_t ts_width = timestamp_width(time_precision);
 
 			fprintf(out_fd, "%s[%*.*f] (%*.*f)%s ",
 				use_colors ? KGRN : "",
